mte/requirements.txt
paramiko==3.1.0
virtualbox==2.1.1
GitPython==3.1.31
PyYAML==5.4.1